Divide FastSync into smaller parts: - create FastSyncBlockBodiesHandler - create FastSyncBlockHeadersHandler - rename BlocksValidator to FastSyncBlockValidator - create FastSyncFixtures - rename HandlerState to FastSyncHandlerState and move it to separate file - rename SyncState to FastSyncState and move it to separate file - create FastSyncNodeHandler - create FastSyncReceiptsHandler
